@@ -17,76 +17,76 @@
   },
   "homepage": "https://github.com/artem-malko/react-ssr-template#readme",
   "dependencies": {
-    "@types/cookie-parser": "1.4.2",
+    "@types/cookie-parser": "1.4.3",
     "@types/express": "4.17.13",
     "@types/express-useragent": "1.0.2",
     "@types/inline-style-prefixer": "5.0.0",
-    "@types/js-cookie": "3.0.1",
+    "@types/js-cookie": "3.0.2",
     "@types/loader-utils": "2.0.3",
-    "@types/node": "17.0.24",
+    "@types/node": "17.0.31",
     "@types/pino": "7.0.5",
-    "@types/react": "18.0.5",
-    "@types/react-dom": "18.0.1",
+    "@types/react": "18.0.9",
+    "@types/react-dom": "18.0.3",
     "@types/react-redux": "7.1.24",
     "@types/redux-logger": "3.0.9",
     "@types/uuid": "8.3.4",
     "@types/webpack-env": "1.16.4",
-    "axios": "0.26.1",
+    "axios": "0.27.2",
     "bidi-css-js": "3.1.0",
     "compression-webpack-plugin": "9.2.0",
     "cookie-parser": "1.4.6",
-    "esbuild": "0.14.36",
+    "esbuild": "0.14.38",
     "esbuild-loader": "2.18.0",
-    "express": "4.17.3",
+    "express": "4.18.1",
     "express-useragent": "1.0.15",
     "inline-style-prefixer": "6.0.1",
     "js-cookie": "3.0.1",
     "loader-utils": "3.2.0",
     "lodash.mergewith": "4.6.2",
-    "path-to-regexp": "6.2.0",
-    "pino": "7.10.0",
+    "path-to-regexp": "6.2.1",
+    "pino": "7.11.0",
     "pino-pretty": "7.6.1",
-    "react": "18.0.0",
-    "react-dom": "18.0.0",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
     "react-error-boundary": "3.1.4",
-    "react-query": "4.0.0-beta.3",
-    "react-redux": "8.0.0",
-    "redux": "4.1.2",
+    "react-query": "4.0.0-beta.7",
+    "react-redux": "8.0.1",
+    "redux": "4.2.0",
     "redux-logger": "3.0.6",
     "set-cookie-parser": "2.4.8",
     "terser-webpack-plugin": "5.3.1",
     "ts-node": "10.7.0",
-    "tslib": "2.3.1",
-    "typescript": "4.6.3",
+    "tslib": "2.4.0",
+    "typescript": "4.6.4",
     "uuid": "8.3.2",
     "webpack": "5.72.0",
     "webpack-cli": "4.9.2",
     "webpack-stats-plugin": "1.0.3"
   },
   "devDependencies": {
-    "@testing-library/react": "13.1.1",
+    "@testing-library/react": "13.2.0",
     "@types/chai": "4.3.1",
     "@types/chai-dom": "0.0.12",
-    "@types/mocha": "9.1.0",
+    "@types/mocha": "9.1.1",
     "@types/sinon": "10.0.11",
-    "@typescript-eslint/eslint-plugin": "5.19.0",
-    "@typescript-eslint/parser": "5.19.0",
+    "@typescript-eslint/eslint-plugin": "5.22.0",
+    "@typescript-eslint/parser": "5.22.0",
     "axios-mock-adapter": "1.20.0",
     "caniuse-api": "3.0.0",
     "chai": "4.3.6",
     "chai-dom": "1.11.0",
-    "eslint": "8.13.0",
+    "eslint": "8.15.0",
     "eslint-plugin-functional": "4.2.1",
     "eslint-plugin-react": "7.29.4",
-    "eslint-plugin-react-hooks": "4.4.0",
+    "eslint-plugin-react-hooks": "4.5.0",
     "jsdom": "19.0.0",
-    "mocha": "9.2.2",
-    "mock-local-storage": "1.1.21",
+    "mocha": "10.0.0",
+    "mock-local-storage": "1.1.22",
     "morgan": "1.10.0",
-    "nodemon": "2.0.15",
+    "nodemon": "2.0.16",
     "nyc": "15.1.0",
     "prettier": "2.6.2",
-    "sinon": "13.0.2",
+    "sinon": "14.0.0",
     "source-map-support": "0.5.21",
     "webpack-bundle-analyzer": "4.5.0",
     "webpack-npm-dependencies-analyzer": "1.0.1"
