@@ -4,6 +4,7 @@ import { AnyPage, URLQueryParams } from 'infrastructure/router/types';
 import { NewsPage } from 'ui/pages/news';
 import { NewsItemPage } from 'ui/pages/newsItem';
 import { HttpErrorCode } from 'core/types/http';
+import { UsersPage } from 'ui/pages/users';
 
 export type AppState = {
   appContext: AppContext;
@@ -15,8 +16,9 @@ export interface AppContext {
 }
 
 export interface CommonPage extends AnyPage<string> {
+  // @TODO maybe remove it?
   errorCode?: HttpErrorCode;
 }
 
-export type Page = RootPage | ErrorPage | NewsPage | NewsItemPage;
+export type Page = RootPage | ErrorPage | NewsPage | NewsItemPage | UsersPage;
 export type PageName = Page['name'];
