@@ -2,7 +2,8 @@ import { createActionConstructor } from 'core/actions/createActionConstructor';
 import { Popup } from '../types';
 
 export const showPopupActionType = 'popup_showPopup';
-export const showPopup = createActionConstructor<Popup>(showPopupActionType).createActionCreator();
+export const showPopup =
+  createActionConstructor<Omit<Popup, 'id'>>(showPopupActionType).createActionCreator();
 
 export const hidePopupByIdActionType = 'popup_hidePopupByIdPopup';
 export const hidePopupById = createActionConstructor<{ popupId: string }>(
