@@ -1,4 +1,4 @@
-import { openPage } from 'core/signals/page';
+import { openPageSignal } from 'core/signals/page';
 import { HttpErrorCode } from 'core/types/http';
 import { Route, RouteWithParams } from 'infrastructure/router/types';
 import { ErrorPage } from '.';
@@ -6,7 +6,7 @@ import { ErrorPage } from '.';
 export const errorPageRoute: Route<RouteWithParams<{ code: string }>, ErrorPage> = {
   path: '/error/:code',
   signal: ({ code }) => {
-    return openPage({
+    return openPageSignal({
       name: 'error',
       params: {
         code: parseInt(code, 10) as HttpErrorCode,
